Conflicts:
	final-project/target/m2e-wtp/web-resources/META-INF/maven/com.kh/final3/pom.properties

충돌해

+++ final-project/src/main/java/com/kh/final3/member/controller/MemberController.java
@@ -78,9 +78,16 @@ public class MemberController {
 		String userNo = Integer.toString(userDetails.getUserNo());
 		
 		Member member = memberService.selectMemberByUserNo(userNo);
+		
+		int idx = member.getAddress().indexOf("/");
+		String mainAddress = member.getAddress().substring(0, idx);
+		String subAddress = member.getAddress().substring(idx+1);
+		
 		MemberAttachment memberAttachment = memberService.selectMemberAttachment(userNo);
 		
 		model.addAttribute("member", member);
+		model.addAttribute("mainAddress", mainAddress);
+		model.addAttribute("subAddress", subAddress);
 		model.addAttribute("memberAttachment", memberAttachment);
 		
 		//요청한 requestMapping에 따라 다른 페이지로 보내기
@@ -129,7 +136,8 @@ public class MemberController {
 	}
 	
 	@PostMapping("/update.me")
-	public ModelAndView updateMember(ModelAndView mv, Member member, String detailAddress, MultipartFile upfile, MemberAttachment memberAttachment, HttpSession session) {
+	public ModelAndView updateMember(ModelAndView mv, Member member, String detailAddress, MultipartFile upfile, MemberAttachment memberAttachment, HttpSession session
+			,String deletePhotoCheck) {
 		String check = "none";
 		
 		SaveFile saveFile = new SaveFile();
@@ -161,16 +169,22 @@ public class MemberController {
 			memberAttachment.setChangeName("resources/uploadFiles/memberProfile/"+changeName);
 		}else {
 			MemberAttachment match = memberService.selectMemberAttachment(Integer.toString(member.getUserNo()));
-			if(match != null) {
+			if(match != null && deletePhotoCheck.equals("true")) {
 				new File(session.getServletContext().getRealPath(match.getChangeName())).delete();
-				System.out.println(match.getChangeName());
 				memberAttachment.setOriginName("청록이.jpg");
 				memberAttachment.setChangeName("resources/image/청록이.jpg");
+				
+				Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+				CustomUserDetails userDetails = (CustomUserDetails)principal;
+				memberAttachment.setRefUno(userDetails.getUserNo());
+				
 				memberService.updateMemberAttachment(memberAttachment);
 			}
 		}
 		
-		member.setAddress(member.getAddress() + " " + detailAddress);
+		
+		
+		member.setAddress(member.getAddress() + "/" + detailAddress);
 		
 		memberService.updateMember(member);
 		
@@ -182,6 +196,23 @@ public class MemberController {
 		
 		mv.setViewName("redirect:myPage.me");
 		return mv;
+	}
+	
+	@PostMapping("/delete.me")
+	public String deleteMember(HttpSession session) {
+		
+		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		CustomUserDetails userDetails = (CustomUserDetails)principal;
+		
+		
+		int result = memberService.deleteMember(userDetails.getUserNo());
+		
+		if(result>0) {
+			session.setAttribute("alertMsg", "탈퇴하셨습니다.");
+			SecurityContextHolder.clearContext();
+		}
+		
+		return "redirect:/";
 	}
 	
 	@GetMapping("/memberEnroll.me")
@@ -213,7 +244,7 @@ public class MemberController {
 			member.setAuth("ROLE_MEMBER");
 		}
 		
-		member.setAddress(member.getAddress()+ " " +detailAddress);
+		member.setAddress(member.getAddress()+ "/" +detailAddress);
 		member.setUserPwd("1234");
 		
 		int result = memberService.insertMember(member);
@@ -255,8 +286,9 @@ public class MemberController {
 		Member member = Member.builder().userPwd(password).userNo(userDetails.getUserNo()).build();
 		int result = memberService.updatePwd(member);
 		if(result>0) {
-			mv.setViewName("redirect:myPage.me");
-			session.setAttribute("alertMsg", "비밀번호 변경완료");
+			mv.setViewName("redirect:/");
+			SecurityContextHolder.clearContext();
+			session.setAttribute("alertMsg", "비밀번호 변경완료. 다시 로그인 해주세요.");
 		}else {
 			
 		}
